Minor changes to code and README
